cli/user-management/requirements.txt
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==5.28.3
setuptools==75.3.0
